requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0

# FastAPI Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Database Dependencies
sqlalchemy>=2.0.0

# HTTP Client
requests>=2.31.0
